@@ -26,12 +26,12 @@
     "classnames": "^2.2.5",
     "lodash": "^4.7.0",
     "moment": "^2.12.0",
-    "react": "^v15.0.0",
+    "prop-types": "^15.7.2",
+    "react": "v15.4.0 ",
     "react-blur-admin": "^0.10.0",
-    "react-dom": "^v15.0.0",
+    "react-dom": "v15.4.0 ",
     "react-flex-proto": "^1.0.0",
-    "react-gl-maps": "git+https://github.com/knledg/react-gl-maps.git",
-    "react-router": "^2.0.1"
+    "react-router-dom": "^5.0.1"
   },
   "devDependencies": {
     "babel-core": "^6.7.4",
@@ -59,7 +59,6 @@
     "node-sass": "^3.4.2",
     "postcss-loader": "^0.8.2",
     "raw-loader": "^0.5.1",
-    "react-addons-test-utils": "^v15.0.0",
     "react-transform-catch-errors": "^1.0.2",
     "react-transform-hmr": "^1.0.4",
     "redbox-react": "^1.2.2",
