@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { Page, Panel, Breadcrumbs } from 'react-blur-admin';
-import { Link } from 'react-router';
+import { Link } from 'react-router-dom';
 
 export class NotFound extends React.Component {
 
