@@ -1,6 +1,6 @@
 import React from 'react';
 import { Page, Tabs, Tab, Breadcrumbs } from 'react-blur-admin';
-import { Link } from 'react-router';
+import { Link } from 'react-router-dom';
 
 export class TabsDemo extends React.Component {
 
